+++ app/models.py
@@ -1,0 +1,8 @@
+import strawberry
+
+
+@strawberry.type
+class Memo:
+    id: int
+    title: str
+    description: str
