+++ requirements.txt
@@ -1,39 +1,39 @@
 annotated-types==0.6.0
-anyio==4.3.0
+anyio~=4.3.0
 astroid==3.1.0
 certifi==2024.2.2
 charset-normalizer==3.3.2
-click==8.1.7
+click~=8.1.7
 coverage==6.5.0
 coveralls==3.3.1
 dill==0.3.8
 docopt==0.6.2
-fastapi==0.110.2
+fastapi~=0.110.2
 graphql-core==3.2.3
-h11==0.14.0
+h11~=0.14.0
 httpcore==1.0.5
-httpx==0.27.0
+httpx~=0.27.0
 idna==3.7
 iniconfig==2.0.0
 isort==5.13.2
-mangum==0.17.0
+mangum~=0.17.0
 mccabe==0.7.0
 packaging==24.0
 platformdirs==4.2.1
 pluggy==1.5.0
-pydantic==2.7.1
-pydantic_core==2.18.2
+pydantic~=2.7.1
+pydantic_core~=2.18.2
 pylint==3.1.0
-pytest==8.1.1
+pytest~=8.1.1
 pytest-cov==5.0.0
-python-dateutil==2.9.0.post0
-python-multipart==0.0.9
-requests==2.31.0
-six==1.16.0
-sniffio==1.3.1
-starlette==0.37.2
+python-dateutil~=2.9.0.post0
+python-multipart~=0.0.9
+requests~=2.31.0
+six~=1.16.0
+sniffio~=1.3.1
+starlette~=0.37.2
 strawberry-graphql==0.227.2
 tomlkit==0.12.4
-typing_extensions==4.11.0
+typing_extensions~=4.11.0
 urllib3==2.2.1
-uvicorn==0.29.0
+uvicorn~=0.29.0
